+++ robot/src/main/include/Manipulator.h
@@ -11,9 +11,9 @@ class Manipulator
         Manipulator(Intake *intake, Feeder *feeder, Shooter *shooter);
         int Round();
         void PrepareShot(int rpm, double aimHeight);
-        int GetSelectedRPM(int index);
-        double GetSelectedHoodPosition(int index);
-        void ContinuousShoot(int aim_position, double geneva_speed, int shooting_rpm);
+        int GetSelectedRPM(unsigned int index);
+        double GetSelectedHoodPosition(unsigned int index);
+        void ContinuousShoot(double aim_position, double geneva_speed, int shooting_rpm);
         void ContinuousIntake();
         void ResetManipulatorElements();
         void GetState();
